Dev1

+++ src/com/endro32/proadmin/util/FileManager.java
@@ -23,22 +23,31 @@ public class FileManager {
 	
 	public static boolean mkdir(String path) {
 		File dir = new File(appdir+"/"+path);
-		return dir.mkdir();
+		if(!dir.exists()) {
+			System.out.println("Creating directory "+appdir+"/"+path);
+			return dir.mkdir();
+		} else return false;
 	}
 	
 	public static boolean delFile(String path) {
 		File file = new File(appdir+"/"+path);
-		return file.delete();
+		if(file.exists() && !file.isDirectory()) {
+			System.out.println("Attempting to delete file "+appdir+"/"+path);
+			return file.delete();
+		} else return false;
 	}
 	
 	public static boolean delDir(String path) {
 		File dir = new File(appdir+"/"+path);
-		try {
-			FileUtils.deleteDirectory(dir);
-		} catch (IOException e) {
-			return false;
-		}
-		return true;
+		if(dir.exists() && dir.isDirectory()) {
+			System.out.println("Attempting to delete directory "+appdir+"/"+path);
+			try {
+				FileUtils.deleteDirectory(dir);
+			} catch (IOException e) {
+				return false;
+			}
+			return true;
+		} else return false;
 	}
 	
 	public static boolean updateFileSystem() {
@@ -62,35 +71,13 @@ public class FileManager {
 			mkdir("groups/"+group);
 			if(Config.getMode(group).equals(GroupMode.INDIVIDUAL)) {
 				for(String name : Config.getServerNamesForGroup(group)) {
-					mkdir("groups/"+group+"/"+name);
-					mkdir("groups/"+group+"/"+name+"/plugins");
-					File eula = new File("groups/"+group+"/"+name+"/eula.txt");
-					if(!eula.exists()) {
-						FileGenerator.generateEULA("groups/"+group+"/"+name, 
-								Config.getEULAStatus());
-					} else if(eula.exists()){
-						FileGenerator.updateEULA("groups/"+group+"/"+name, 
-								Config.getEULAStatus());
-					}
-					extractResource("server.properties", "groups/"+group+"/"+name+"/server.properties",
-							false);
+					updateServerDirectory(group, name);
 				}
 			} else if(Config.getMode(group).equals(GroupMode.CLONED)) {
 				List<String> servers = Config.getServerNamesForGroup(group);
 				String first = servers.get(0);
-				mkdir("groups/"+group+"/"+first);
-				mkdir("groups/"+group+"/"+first+"/plugins");
-				File eula = new File("groups/"+group+"/"+first+"/eula.txt");
-				if(!eula.exists()) {
-					FileGenerator.generateEULA("groups/"+group+"/"+first,
-							Config.getEULAStatus());
-				} else if(eula.exists()){
-					FileGenerator.updateEULA("groups/"+group+"/"+first, 
-							Config.getEULAStatus());
-				}
-				extractResource("server.properties", "groups/"+group+"/"+first+"/server.properties",
-						false);
-				File firstDir = eula.getParentFile();
+				updateServerDirectory(group, first);
+				File firstDir = new File(appdir+"/groups/"+group+"/"+first);
 				for(String name : servers) {
 					File target = new File(appdir+"/groups/"+group+"/"+name);
 					if(name.equals(first)) continue;
@@ -105,8 +92,32 @@ public class FileManager {
 		return true;
 	}
 	
+	public static boolean updateServerDirectory(String group, String server) {
+		try {
+			System.out.println("Attempting to update "+server+" in "+group+"...");
+			mkdir("groups/"+group+"/"+server);
+			mkdir("groups/"+group+"/"+server+"/plugins");
+			File eula = new File("groups/"+group+"/"+server+"/eula.txt");
+			if(!eula.exists()) {
+				FileGenerator.generateEULA("groups/"+group+"/"+server, 
+						Config.getEULAStatus());
+			} else if(eula.exists()){
+				FileGenerator.updateEULA("groups/"+group+"/"+server, 
+						Config.getEULAStatus());
+			}
+			extractResource("server.properties", "groups/"+group+"/"+server+"/server.properties",
+					false);
+			System.out.println("Successfully updated "+server);
+			return true;
+		} catch(Exception e) {
+			System.out.println("Failed to update "+server);
+			return false;
+		}
+	}
+	
 	public static boolean download(URL url, String path, boolean replaceExisting) {
 		try {
+			System.out.println("Attempting to download "+url+"...");
 			ReadableByteChannel rbc = Channels.newChannel(url.openStream());
 			FileOutputStream fos = new FileOutputStream("tmp");
 			fos.getChannel().transferFrom(rbc,  0, Long.MAX_VALUE);
@@ -118,13 +129,16 @@ public class FileManager {
 				Files.move(tmp.toPath(), target.toPath());
 			}
 			fos.close();
+			System.out.println("Successfully downloaded to "+appdir+"/"+path);
 		} catch (IOException e) {
+			System.out.println("Failed to download "+url);
 			return false;
 		}
 		return true;
 	}
 	
 	public static boolean extractResource(String name, String target, boolean replaceExisting) {
+		System.out.println("Attempting to extract "+name+"...");
 		InputStream link = Main.class.getResourceAsStream("resources/"+name);
 		File file = new File(appdir+"/"+target);
 		try {
@@ -133,7 +147,9 @@ public class FileManager {
 			} else {
 				Files.copy(link, file.toPath());
 			}
+			System.out.println("Succesfully extracted to "+appdir+"/"+target);
 		} catch(IOException e) {
+			System.out.println("Failed to extract "+name);
 			return false;
 		}
 		return true;
